moving helper functions to helper.ts
